docs: add RSC example to demo app
