+++ apps/demo/config/blocks/Hero/Hero.tsx
@@ -1,0 +1,97 @@
+/* eslint-disable @next/next/no-img-element */
+import React from "react";
+import styles from "./styles.module.css";
+import { getClassNameFactory } from "@/core/lib";
+import { Button } from "@/core/components/Button";
+import { Section } from "../../components/Section";
+import { PuckComponent } from "@/core/types";
+
+const getClassName = getClassNameFactory("Hero", styles);
+
+export type HeroProps = {
+  quote?: { index: number; label: string };
+  title: string;
+  description: string;
+  align?: string;
+  padding: string;
+  image?: {
+    mode?: "inline" | "background";
+    url?: string;
+  };
+  buttons: {
+    label: string;
+    href: string;
+    variant?: "primary" | "secondary";
+  }[];
+};
+
+export const Hero: PuckComponent<HeroProps> = ({
+  align,
+  title,
+  description,
+  buttons,
+  padding,
+  image,
+  puck,
+}) => {
+  return (
+    <Section
+      className={getClassName({
+        left: align === "left",
+        center: align === "center",
+        hasImageBackground: image?.mode === "background",
+      })}
+      style={{ paddingTop: padding, paddingBottom: padding }}
+    >
+      {image?.mode === "background" && (
+        <>
+          <div
+            className={getClassName("image")}
+            style={{
+              backgroundImage: `url("${image?.url}")`,
+            }}
+          ></div>
+
+          <div className={getClassName("imageOverlay")}></div>
+        </>
+      )}
+
+      <div className={getClassName("inner")}>
+        <div className={getClassName("content")}>
+          <h1>{title}</h1>
+          <p className={getClassName("subtitle")}>{description}</p>
+          <div className={getClassName("actions")}>
+            {buttons.map((button, i) => (
+              <Button
+                key={i}
+                href={button.href}
+                variant={button.variant}
+                size="large"
+                tabIndex={puck.isEditing ? -1 : undefined}
+              >
+                {button.label}
+              </Button>
+            ))}
+          </div>
+        </div>
+
+        {align !== "center" && image?.mode !== "background" && image?.url && (
+          <div
+            style={{
+              backgroundImage: `url('${image?.url}')`,
+              backgroundSize: "cover",
+              backgroundRepeat: "no-repeat",
+              backgroundPosition: "center",
+              borderRadius: 24,
+              height: 356,
+              marginLeft: "auto",
+              width: "100%",
+            }}
+          />
+        )}
+      </div>
+    </Section>
+  );
+};
+
+export default Hero;
