+++ apps/demo/config/blocks/Template/Template.tsx
@@ -1,0 +1,25 @@
+import React from "react";
+import { Slot } from "@/core/types";
+import styles from "./styles.module.css";
+import { getClassNameFactory } from "@/core/lib";
+import { Section } from "../../components/Section";
+import { PuckComponent } from "@/core/types";
+
+const getClassName = getClassNameFactory("Template", styles);
+
+export type TemplateProps = {
+  template: string;
+  children: Slot;
+};
+
+export const Template: PuckComponent<TemplateProps> = ({
+  children: Children,
+}) => {
+  return (
+    <Section>
+      <Children className={getClassName()} />
+    </Section>
+  );
+};
+
+export default Template;
