+++ apps/demo/config/components/Header/index.tsx
@@ -5,7 +5,10 @@ import styles from "./styles.module.css";
 const getClassName = getClassNameFactory("Header", styles);
 
 const NavItem = ({ label, href }: { label: string; href: string }) => {
-  const navPath = window.location.pathname.replace("/edit", "") || "/";
+  const navPath =
+    typeof window !== "undefined"
+      ? window.location.pathname.replace("/edit", "") || "/"
+      : "/";
 
   const isActive = navPath === (href.replace("/edit", "") || "/");
