+++ apps/demo/config/root.tsx
@@ -1,6 +1,6 @@
-import { DefaultRootProps, DropZone, RootConfig } from "@/core";
-import { Footer } from "./components/Footer";
+import { DefaultRootProps, RootConfig } from "@/core";
 import { Header } from "./components/Header";
+import { Footer } from "./components/Footer";
 
 export type RootProps = DefaultRootProps;
 
@@ -8,13 +8,14 @@ export const Root: RootConfig<RootProps> = {
   defaultProps: {
     title: "My Page",
   },
-  render: ({ puck }) => {
+  render: ({ puck: { isEditing, renderDropZone: DropZone } }) => {
     return (
       <div
         style={{ display: "flex", flexDirection: "column", minHeight: "100vh" }}
       >
-        <Header editMode={puck.isEditing} />
+        <Header editMode={isEditing} />
         <DropZone zone="default-zone" style={{ flexGrow: 1 }} />
+
         <Footer>
           <Footer.List title="Section">
             <Footer.Link href="#">Label</Footer.Link>
