+++ apps/demo/config/server.tsx
@@ -1,21 +1,19 @@
 import { Button } from "./blocks/Button";
 import { Card } from "./blocks/Card";
 import { Grid } from "./blocks/Grid";
-import { Hero } from "./blocks/Hero";
+import { Hero } from "./blocks/Hero/server";
 import { Heading } from "./blocks/Heading";
 import { Flex } from "./blocks/Flex";
 import { Logos } from "./blocks/Logos";
 import { Stats } from "./blocks/Stats";
-import { Template } from "./blocks/Template";
+import { Template } from "./blocks/Template/server";
 import { Text } from "./blocks/Text";
 import { Space } from "./blocks/Space";
-
 import Root from "./root";
 import { UserConfig } from "./types";
-import { initialData } from "./initial-data";
 
 // We avoid the name config as next gets confused
-export const conf: UserConfig = {
+const conf: UserConfig = {
   root: Root,
   categories: {
     layout: {
@@ -47,9 +45,5 @@ export const conf: UserConfig = {
     Space,
   },
 };
-
-export const componentKey = Buffer.from(
-  `${Object.keys(conf.components).join("-")}-${JSON.stringify(initialData)}`
-).toString("base64");
 
 export default conf;
