+++ apps/demo/config/types.ts
@@ -1,0 +1,38 @@
+import { Config, Data } from "@/core";
+import { ButtonProps } from "./blocks/Button";
+import { CardProps } from "./blocks/Card";
+import { GridProps } from "./blocks/Grid";
+import { HeroProps } from "./blocks/Hero";
+import { HeadingProps } from "./blocks/Heading";
+import { FlexProps } from "./blocks/Flex";
+import { LogosProps } from "./blocks/Logos";
+import { StatsProps } from "./blocks/Stats";
+import { TemplateProps } from "./blocks/Template";
+import { TextProps } from "./blocks/Text";
+import { SpaceProps } from "./blocks/Space";
+
+import { RootProps } from "./root";
+
+export type { RootProps } from "./root";
+
+export type Props = {
+  Button: ButtonProps;
+  Card: CardProps;
+  Grid: GridProps;
+  Hero: HeroProps;
+  Heading: HeadingProps;
+  Flex: FlexProps;
+  Logos: LogosProps;
+  Stats: StatsProps;
+  Template: TemplateProps;
+  Text: TextProps;
+  Space: SpaceProps;
+};
+
+export type UserConfig = Config<
+  Props,
+  RootProps,
+  "layout" | "typography" | "interactive"
+>;
+
+export type UserData = Data<Props, RootProps>;
